ISSUE-30: Delete icon file
Better names for API test case files
