+++ backend/integration-tests/api/icon-delete.spec.ts
